♻️ Refactor: 스터디 그룹 API 요청/응답 타입 api/types로 이동 (#203)
