+++ src/api/types/index.ts
@@ -1,1 +1,2 @@
 export * from '@api/types/auth'
+export * from '@api/types/group'
