+++ src/hooks/queries/useStudyGroupMutations.ts
@@ -3,7 +3,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { studyApi } from '@api'
 import { studyQueryKey } from '@hooks'
 
-import type { CreateStudyGroupResponse } from '@models'
+import type { CreateStudyGroupResponse } from '@api'
 import type { AxiosResponse } from 'axios'
 
 export const useStudyGroupMutations = () => {
