--- src/types/api-request-types/index.ts
@@ -1,1 +1,0 @@
-export * from '@models/api-request-types/studyGroup-request-types'
