--- src/types/api-response-types/index.ts
@@ -1,1 +1,0 @@
-export * from '@models/api-response-types/studyGroup-response-types'
