--- src/types/api-response-types/studyGroup-response-types.ts
@@ -1,11 +1,0 @@
-export interface CreateStudyGroupResponse {
-  uuid: string
-  name: string
-  introduction: string
-  maxHeadcount: number
-  profileImgUrl: string
-  startAt: string
-  endAt: string
-  lectures: number[]
-  createdAt: string
-}
