+++ src/types/index.ts
@@ -5,5 +5,3 @@ export * from '@models/review'
 export * from '@models/lecture'
 export * from '@models/chat'
 export * from '@models/user'
-export * from '@models/api-request-types'
-export * from '@models/api-response-types'
